mpi4py: run the spawn and dynamic process tests
Split the mpi4py Github Action into 4 parts:

1. build: do everything to build, configure, and install Open MPI and
   mpi4py
2. run: run all the mpi4py tests with its defaults.  As of March 2024,
   this disables the spawn and dynamic tests, which means that the
   entire block of tests should pass.
3. run_spawn: run all the mpi4py tests, including the spawn tests, but
   only if the "mpi4py" label is set on the PR.  As of March 2024, we
   know some of these tests are failing.
4. run_dynamic: run all the mpi4py tests, including the dynamic tests,
   but only if the "mpi4py" label is set on the PR.  As of March 2024,
   we know some of these tests are failing.

The spawn and dynamic failures are different, so we split them up and
run them separately.

Note that for steps 2, 3, and 4, we use a GitHub reusable workflow so
that we don't have to duplicate the code -- we just set a different
environment variable for each, which triggers different behavior.

Signed-off-by: Jeff Squyres <jeff@squyres.com>